+++ CHANGELOG.md
@@ -60,6 +60,7 @@
 - New #73: Add `editMessageCaption` method.
 - New #74: Add `editMessageMedia` method and `MessageEntityType` class with constants.
 - New #75: Add `editMessageLiveLocation` and `stopMessageLiveLocation` methods.
+- New #76: Add `editMessageReplyMarkup` method.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.
 - Chg #30: Remove `TelegramRequestWithFilesInterface`.